Upload de Projeto Prático | Leads API (Repositórios para Campanhas e seus Leads) - Módulo SQL + Node.js

+++ Backend/Projetos Práticos/leads-api/src/repositories/prisma/PrismaCampaignsRepository.ts
@@ -1,5 +1,5 @@
 import { Campaign } from "@prisma/client";
-import { CampaignsRepository, CreateCampaignAttributes } from "../campaigns-repository";
+import { AddLeadToCampaignAttributes, CampaignsRepository, CreateCampaignAttributes } from "../campaigns-repository";
 import { prisma } from "../../database";
 
 export class PrismaCampaignsRepository implements CampaignsRepository {
@@ -10,7 +10,7 @@ export class PrismaCampaignsRepository implements CampaignsRepository {
     findById(campaignId: number): Promise<Campaign | null> {
         return prisma.campaign.findUnique({
             where: { id: campaignId },
-            include: { leads: true }
+            include: { leads: { include: { lead: true } } }
         });
     };
 
@@ -20,17 +20,48 @@ export class PrismaCampaignsRepository implements CampaignsRepository {
         });
     };
 
-    updateById(campaignId: number, attributes: Partial<CreateCampaignAttributes>): Promise<Campaign | null> {
+    async updateById(campaignId: number, attributes: Partial<CreateCampaignAttributes>): Promise<Campaign | null> {
+        const campaignExists = await prisma.campaign.findUnique({ where: { id: campaignId } });
+        if (!campaignExists) return null;
+
         return prisma.campaign.update({
             where: { id: campaignId },
             data: attributes
         });
     };
 
-    deleteById(campaignId: number): Promise<Campaign | null> {
+    async deleteById(campaignId: number): Promise<Campaign | null> {
+        const campaignExists = await prisma.campaign.findUnique({ where: { id: campaignId } });
+        if (!campaignExists) return null;
+
         return prisma.campaign.delete({
             where: { id: campaignId }
         });
     };
 
+    async addLead(attributes: AddLeadToCampaignAttributes): Promise<void> {
+        await prisma.leadCampaign.create({ data: attributes })
+    }
+
+    async updateLeadStatus(attributes: AddLeadToCampaignAttributes): Promise<void> {
+        await prisma.leadCampaign.update({
+            data: { status: attributes.status },
+            where: {
+                leadId_campaignId: {
+                    campaignId: attributes.campaignId,
+                    leadId: attributes.leadId
+                }
+            }
+        });
+    };
+
+    async removeLead(campaignId: number, leadId: number): Promise<void> {
+        await prisma.leadCampaign.delete({
+            where: {
+                leadId_campaignId: { campaignId, leadId }
+            }
+        });
+    };
+
 }
+
